requirements.txt
aniso8601==8.0.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-RESTful==0.3.7
google-api-core==1.15.0
google-auth==1.10.0
google-cloud-bigquery==1.23.1
google-cloud-core==1.1.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.4
pandas==0.25.3
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
StackAPI==0.1.12
urllib3==1.25.7
visitor==0.1.3
Werkzeug==0.16.0
